Fixes lit test that doesn't fail on windows
long in linux 64 is 64 bits but is always 32 bits on windows.  The lit test was modified 
to use long long instead of long and check for 64-bit mangling.

llvm-svn: 193901

+++ clang/test/CodeGenCXX/builtins.cpp
@@ -17,13 +17,13 @@ S *addressof(bool b, S &s, S &t) {
 }
 
 extern "C" int __builtin_abs(int); // #1
-long __builtin_abs(long);          // #2
+long long __builtin_abs(long long);          // #2
 extern "C" int __builtin_abs(int); // #3
 
 int x = __builtin_abs(-2);
 // CHECK: entry:
 // CHECK-NEXT:  store i32 2, i32* @x, align 4
 
-long y = __builtin_abs(-2l);
+long y = __builtin_abs(-2ll);
 // CHECK: entry:
-// CHECK-NEXT:  %call = call i32 @_Z13__builtin_absl(i32 -2)
+// CHECK-NEXT:  %call = call i64 @_Z13__builtin_absx(i64 -2)
